pyproject.toml
[tool.poetry]
name = "hack2020-final"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
pyopenssl = "^19.1.0"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
